% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASS_package.R
\name{KfoldTcompute}
\alias{KfoldTcompute}
\title{Estimating the local FDR of unlabeled individuals under PASSC}
\usage{
KfoldTcompute(
  X_label,
  Y_label,
  X_test,
  Null_label,
  W_test_whole,
  K,
  lambda,
  algo,
  classification = FALSE
)
}
\arguments{
\item{X_label}{Matrix. The covariate matrix of the labeled data.}

\item{Y_label}{Vector. Responses of the labeled data.}

\item{X_test}{Matrix. The covariate matrix of the unlabeled data.}

\item{Null_label}{Vector. Indices of the individuals not of interest among the labeled data.}

\item{W_test_whole}{Vector. Prediction values on the unlabeled data given by the model trained on the whole labeled data set without splitting.}

\item{K}{Integral. Number of folds used for PASSC. Five as default.}

\item{lambda}{Numeric. The tuning parameter used for fitting the predictive model.}

\item{algo}{Class. A class specifying the algorithms used for fitting and prediction. The class should be well-defined with two functions. One named "fitting", which takes algorithm class, covariate matrix, responses and possibly a tuning parameter as inputs and outputs a model object. The other named "Pred", which takes algorithm class, a model object and a new covariate matrix as inputs and outputs a vector of predciton values.}

\item{classification}{classification Logical. If the response is discrete then set this to TRUE. FALSE as default.}
}
\value{
Vector containing the estimated local FDR of the unlabeled individuals under PASSC.
}
\description{
Estimating the local FDR of unlabeled individuals under PASSC
}
