% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASS_package.R
\name{PASS_sampling}
\alias{PASS_sampling}
\title{Implementation of the PASS subsampling method}
\usage{
PASS_sampling(
  X_label,
  Y_label,
  X_test,
  Null_label,
  g = "RBF",
  m,
  alpha,
  algo,
  cv = FALSE,
  K = 5,
  lambda = 500,
  classification = FALSE
)
}
\arguments{
\item{X_label}{Matrix. The covariate matrix of the labeled data.}

\item{Y_label}{Vector. Responses of the labeled data.}

\item{X_test}{Matrix. The covariate matrix of the unlabeled data.}

\item{Null_label}{Vector. Indices of the individuals not of interest among the labeled data.}

\item{g}{Character. The criterion function used for measuring similarity. Take "RBF" or "COS" and "RBF" as default.}

\item{m}{Integral. The sampling budget, that is, how many unlabeled individuals should be sampled.}

\item{alpha}{Numeric. The pre-specified error rate level.}

\item{algo}{Class. A class specifying the algorithms used for fitting and prediction. The class should be well-defined with two functions. One named "fitting", which takes algorithm class, covariate matrix, responses and possibly a tuning parameter as inputs and outputs a model object. The other named "Pred", which takes algorithm class, a model object and a new covariate matrix as inputs and outputs a vector of predciton values.}

\item{cv}{Logical. Set to TRUE if PASSC is to be implemented. FALSE as default.}

\item{K}{Integral. Number of folds used for PASSC. Five as default.}

\item{lambda}{Numeric. The tuning parameter used for fitting the predictive model.}

\item{classification}{Logical. If the response is discrete then set this to TRUE. FALSE as default.}
}
\value{
A list containing a vector of sampling probabilities supported on the unlabeled data set and a vector indicating which individuals are sampled.
}
\description{
Implementation of the PASS subsampling method
}
