% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASS_package.R
\name{SolvePi}
\alias{SolvePi}
\title{Solving the quadratic optimization problem for the sampling probability vector}
\usage{
SolvePi(TN, m, S, Screened, alpha)
}
\arguments{
\item{TN}{Vector. Estimated local FDR of the unlabeled individuals.}

\item{m}{Integral. The sampling budget, that is, how many unlabeled individuals should be sampled.}

\item{S}{Matrix. The similarity matrix of the unlabeled covariates.}

\item{Screened}{Vector. Indices of the screened individuals.}

\item{alpha}{Numeric. The pre-specified error rate level.}
}
\value{
The subsampling probability vector.
}
\description{
Solving the quadratic optimization problem for the sampling probability vector
}
