% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASS_package.R
\name{Tcompute}
\alias{Tcompute}
\title{Estimating the local FDR of unlabeled individuals}
\usage{
Tcompute(W_cal, W_test, Null_cal, classification = FALSE)
}
\arguments{
\item{W_cal}{Vector. Prediction values on the calibration data given by the model trained on the training set.}

\item{W_test}{Vector. Prediction values on the unlabeled data given by the model trained on the training set.}

\item{Null_cal}{Vector. Indices of the individuals not of interest among the calibration data.}

\item{classification}{Logical. Should be TRUE if under classification settings. FALSE as default.}
}
\value{
Vector containing the estimated local FDR of the unlabeled individuals.
}
\description{
Estimating the local FDR of unlabeled individuals
}
